+++ src/main/java/Duke.java
@@ -1,3 +1,10 @@
+import command.Command;
+import others.DukeException;
+import parser.Parser;
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 import java.io.*;
 
 public class Duke {
@@ -16,8 +23,8 @@ public class Duke {
             ui.print("No existing To Do List. Generating a new one :)");
             tasks = new TaskList();
         }
-        catch (DukeException e) {
-            ui.showError(e.getMessage());
+        catch (IOException e) {
+            ui.print("Unable to read specific lines in data.txt file");
         }
     }
 
@@ -38,20 +45,6 @@ public class Duke {
             } finally {
                 ui.showLine();
             }
-
-            /*
-            try {
-                String fullCommand = ui.readCommand();
-                ui.showLine(); // show the divider line ("_______")
-                Command c = Parser.parse(fullCommand);
-                c.execute(tasks, ui, storage);
-                isExit = c.isExit();
-            } catch (DukeException e) {
-                ui.showError(e.getMessage());
-            } finally {
-                ui.showLine();
-            }
-             */
         }
     }
