--- src/main/java/TaskList.java
@@ -1,52 +1,0 @@
-import java.util.ArrayList;
-import java.io.*;
-
-public class TaskList {
-    protected static ArrayList<Task> taskList;
-
-    public TaskList(String input) throws DukeException {
-        try {
-            this.taskList = new ArrayList();
-            BufferedReader reader = new BufferedReader(new StringReader(input));
-            String line = null;
-            while ((line = reader.readLine()) != null) {
-                String[] delimited = line.split(";");
-                if ("T".equals(delimited[1])) {
-                    taskList.add(new ToDo(delimited[3]));
-                } else if ("E".equals(delimited[1])) {
-                    taskList.add(new Event(delimited[3], delimited[4]));
-                } else if ("D".equals(delimited[1])) {
-                    taskList.add(new Deadline(delimited[3], delimited[4]));
-                }
-                taskList.get(taskList.size() - 1).setStatus(Boolean.parseBoolean(delimited[2]));
-            }
-        }
-        catch(IOException e) {
-            System.out.println("No data in data.txt file");
-        }
-    }
-
-    public TaskList() {
-        this.taskList = new ArrayList();
-    }
-
-    public boolean isEmpty() {
-        return taskList.isEmpty();
-    }
-
-    public Task get(int i) {
-        return taskList.get(i);
-    }
-
-    public int size() {
-        return taskList.size();
-    }
-
-    public void remove(int i) {
-        taskList.remove(i);
-    }
-
-    public void add(Task task) {
-        taskList.add(task);
-    }
-}
