+++ src/main/java/command/Command.java
@@ -1,3 +1,10 @@
+package command;
+
+import task.TaskList;
+import ui.Ui;
+import storage.Storage;
+import others.DukeException;
+
 import java.io.IOException;
 
 public abstract class Command {
