+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import task.TaskList;
+import ui.Ui;
+import storage.Storage;
+
 public class ExitCommand extends Command {
     public ExitCommand() {
     }
