+++ src/main/java/command/UnknownCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import task.TaskList;
+import ui.Ui;
+import storage.Storage;
+
 public class UnknownCommand extends Command {
     public UnknownCommand() {}
