+++ src/main/java/others/DukeException.java
@@ -1,5 +1,4 @@
-import java.util.*;
-import java.io.IOException;
+package others;
 
 public class DukeException extends Exception {
     public DukeException(String err) {
