+++ src/main/java/task/Event.java
@@ -1,3 +1,5 @@
+package task;
+
 public class Event extends Task {
     protected String date;
