+++ src/main/java/task/ToDo.java
@@ -1,9 +1,7 @@
-
+package task;
 
 public class ToDo extends Task {
     public ToDo(String desc) {
         super(desc, "T");
     }
-
-
 }
